08_Crewai(Design Patterns)/pyproject.toml
[project]
name = "prompt-chaining"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Abdullah Iqbal", email = "hassangb544@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "crewai>=0.102.0",
    "crewai-tools>=0.36.0",
    "litellm>=1.60.2",
    "python-dotenv>=1.0.1",
]

[project.scripts]
prompt-chaining = "prompt_chaining:main"
simple-chain = "prompt_chaining.simple_chain:kickoff"
simple-plot = "prompt_chaining.simple_chain:plot"
complex-chain = "prompt_chaining.Complex_chain:kickoff"
complex-plot = "prompt_chaining.Complex_chain:plot"



[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
